企画の更新時、本来不要なはずの参加種別の入力欄が必須となっていた問題を修正
